% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronospace_machinery.R
\name{chronospace}
\alias{chronospace}
\title{Create chronospace}
\usage{
chronospace(data_ages, vartype = "non-redundant")
}
\arguments{
\item{data_ages}{A matrix created using \code{\link[=extract_ages]{extract_ages()}}.}

\item{vartype}{Character, indicating the type of variation to be retained ("total" or "non-redundant", see Details;
not meaningful for factors with less than three levels).}
}
\value{
The total and non-redundant percentages of variation accounted for each factor are informed. Also a list
containing the ordination computed for each factor is returned.
}
\description{
Compute the ordination maximizing variation in node ages using between-group PCA (one for each factor).
}
\details{
This function uses between-group PCA to find the set of axes maximizing variation in ages data between the
groups of chronograms obtained through different methodological approaches. By default variation is set as "non-redundant",
meaning bgPCA of each factor is performed using the variation left after removing the portion associated to all the
other factors. If variation = "total" (or if there is only one factor being assessed), bgPCA is performed over the raw
variation in node ages.
}
