% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_outputs.R
\name{plot.chronospace}
\alias{plot.chronospace}
\title{Plot chronospace ordination(s) and axes' extremes}
\usage{
\method{plot}{chronospace}(
  obj,
  output = "all",
  sdev = 1,
  timemarks = NULL,
  gscale = TRUE,
  ellipses = FALSE,
  centroids = FALSE,
  distances = FALSE,
  colors = 1:5,
  factor = 1:(length(obj)),
  axes = c(1, 2),
  pt.alpha = 0.5,
  pt.size = 1.5,
  ell.width = 1.2,
  dist.width = 1,
  ct.size = 5
)
}
\arguments{
\item{obj}{An object of class \code{"chronospace"} containing one or more
ordinations.}

\item{output}{Character, specifying which output should be plotted. Options
are \code{"ordination"} for histograms / scatter plots of bgPC axes,
\code{"extremes"} for trees showing node variation captured by bgPC axes,
\code{"all"} for plotting both outputs, and \code{"none"}.}

\item{sdev}{Numeric, indicating at how many standard deviations should the
extremes of the chronospace axes be depicted. If NULL (the default),
extremes will be depicted at the highest standard deviation avoiding
negative branch lengths.}

\item{timemarks}{Numeric; an optional vector containing ages to be marked by
vertical lines in chronospace representations.}

\item{gscale}{Logical; whether to add chronostratigraphic scale to trees
(via \code{deeptime}).}

\item{ellipses}{Logical, indicating whether to plot data ellipses for each
group (not meaningful for factors with less than three levels).}

\item{centroids}{Logical, indicating whether to plot groups' centroids (not
meaningful for factors with less than three levels).}

\item{distances}{Logical, indicating whether to plot lines between groups'
centroids. The width of the former is inversely proportional to the
distances between the latter in the original variable space (not meaningful
for factors with less than three levels).}

\item{colors}{The colors used to represent groups (i.e. levels) of each
factor.}

\item{factor}{Numeric; the factor or factors whose results are to be
plotted and retained. If \code{NULL}, all the factors are retained.}

\item{axes}{Numeric of length two, specifying the axes to be depicted (not
meaningful for factors with less than three levels).}

\item{pt.alpha}{Numeric, indicating the transparency level of individual
points in the scatter (not meaningful for factors with less than three
levels).}

\item{pt.size}{Numeric, indicating the size of individual points in the
scatter (not meaningful for factors with less than three levels).}

\item{ell.width}{Numeric, indicating line width for data ellipses (not
meaningful for factors with less than three levels).}

\item{dist.width}{Numeric; scaling factor for the width of lines representing
multivariate distances between groups' centroids (not meaningful for
factors with less than three levels).}

\item{ct.size}{Numeric, indicating the size of the points marking groups'
centroids (not meaningful for factors with less than three levels).}
}
\value{
A list containing the histogram/scatterplot and axes' extremes for
each factor included.
}
\description{
For each factor, generate the two basic graphical
representations of a chronospace: the projection of the sampled chronograms
into the ordination axes, and the two 'theoretical' extremes of those axes.
}
\details{
Starting from the object returned by \code{\link[=chronospace]{chronospace()}}, this function
creates the two basic types of plots allowing interpretation of the
ordination maximizing variation in node age between the groups (of
each factor). The first of these is a projection of the phylogenetic trees
(whose ages were used to generate the ordination) into the ordination axes,
and can be either a histogram for two-level factors, or a bivariate
scatter plot for factors with three or more levels. The second output
contains trees representing the positive and negative extremes of each
ordination axis, depicting the variation in nodes ages it captures.
}
\examples{
#Load ages data
data("echinoid_dates")

#Create chronospace
cspace <- chronospace(echinoid_dates)

#Plot chronospace
csp.ord <- plot(cspace)

#Show (bivariate) ordination for factor A
csp.ord$factor_A$ordination

#Show extremes of the first bgPC axis for factor A
csp.ord$factor_A$PC_extremes$bgPC1

#' #Show (univariate) ordination for factor B
csp.ord$factor_B$ordination

#Show extremes of the (only) bgPC axis for factor B
csp.ord$factor_B$PC_extremes$bgPC1
}
