% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronospace_machinery.R
\name{print.chronospace}
\alias{print.chronospace}
\title{Print \code{"dataAges"} objects}
\usage{
\method{print}{chronospace}(x)
}
\arguments{
\item{x}{A \code{"chronospace"} object}
}
\value{
Information on percentages of age variation explained by the included
factors
}
\description{
Print \code{"dataAges"} objects
}
\examples{
#Load ages data
data("data_ages")

#Create chronospace
cspace <- chronospace(data_ages)

#Inspect object
print(cspace)
}
