% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ages.R
\name{print.nodeAges}
\alias{print.nodeAges}
\title{Print \code{"nodeAges"} objects}
\usage{
\method{print}{nodeAges}(x)
}
\arguments{
\item{x}{a \code{"nodeAges"} object.}
}
\value{
Reports the number of trees, factors, and general structure of
the object.
}
\description{
Print \code{"nodeAges"} objects
}
\examples{
data("data_ages")
print(data_ages)
}
