% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_outputs.R
\name{sensitive_nodes}
\alias{sensitive_nodes}
\title{Identify the most sensitive nodes and depict their age distribution}
\usage{
sensitive_nodes(
  data_ages,
  amount_of_change = NULL,
  num_clades = 5,
  factor = 1:ncol(data_ages$factors),
  plot = T,
  colors = 1:5,
  timemarks = NULL,
  gscale = FALSE
)
}
\arguments{
\item{data_ages}{An object of class \code{"nodeAges"}.}

\item{amount_of_change}{Numeric, specifying the desired amount of variation
in node age (expressed in million of years) above which the nodes are
retained and depicted.}

\item{num_clades}{Numeric, indicating the desired number of most sensitive
nodes to be retained and depicted. Ignored if \code{amount_of_change} is
specified.}

\item{factor}{Numeric; the factor or factors whose results are to be
plotted and retained. If \code{NULL}, all the factors are retained.}

\item{plot}{whether to plot the results, or only store them (default=TRUE)}

\item{colors}{The colors used to represent groups (i.e. levels) of each
factor.}

\item{timemarks}{Numeric; an optional vector containing ages to be marked by
vertical lines.}

\item{gscale}{Logical; whether to add chronostratigraphic scale to trees
(via \code{deeptime}).}
}
\value{
A panel showing the relative proportions distribution for the age of
each of the most sensitive nodes associated to each factor.
}
\description{
For each node, identify the most variables nodes (in terms of
age), and plot their age proportion distributions.
}
\details{
This function identifies, for each factor, the nodes in the fixed
topology whose ages are most sensitive (i.e. variable) as a function of the
corresponding levels. These can be done either by indicating a threshold
for variation in age above which nodes are considered relevant, or by
specifying the number of most sensitive nodes which must be retained. The
nodes are named using two terminals selected at random from the two
subclades defined by each node.

For each of these most sensitive nodes, the function will plot the
distribution of relative proportion of ages, by level.
}
\examples{
#Load ages data
data("data_ages")

#Get the 5 most sensitive nodes
sensinodes5 <- sensitive_nodes(data_ages, num_clades = 5, plot = F)

#Show ages distribution for the 5 most sensitive nodes associated to factor A
sensinodes5$factor_A
}
