% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_outputs.R
\name{specified_node}
\alias{specified_node}
\title{Plot the ages of a user-specified node}
\usage{
specified_node(
  data_ages,
  tips = NULL,
  factor = 1:ncol(data_ages$factors),
  plot = TRUE,
  colors = 1:5,
  timemarks = NULL,
  gscale = FALSE
)
}
\arguments{
\item{data_ages}{An object of class \code{"nodeAges"}.}

\item{tips}{Character vector of length 2, specifying the tip names of two
terminals bracketing the node whose age (expressed in million of years) is
to be plotted.}

\item{factor}{Numeric; the factor or factors whose results are to be
plotted and retained. If \code{NULL}, all the factors are retained.}

\item{plot}{whether to plot the results, or only store them (default=TRUE)}

\item{colors}{The colors used to represent groups (i.e. levels) of each
factor.}

\item{timemarks}{Numeric; an optional vector containing ages to be marked by
vertical lines.}

\item{gscale}{Logical; whether to add chronostratigraphic scale to trees
(via \code{deeptime}).}
}
\value{
A panel showing the distribution of ages for the target node under
each level of the requested factors.
}
\description{
Plot the distribution of a given node across conditions.
}
\details{
This function identifies, takes a single character vector containing
the tip names of two terminals, finds the node representing their
most-recent common ancestor, and plots the distribution of inferred ages
for said node across the conditions explored.
}
\examples{
#Load ages data
data("echinoid_dates")

#Get the 5 most sensitive nodes
MRCA_Brissus_Abatus <- specified_node(echinoid_dates, tips = c('Brissus_obesus',
   'Abatus_cordatus'), plot = F)

#Show age distribution for the MRCA of the two terminals associated with
factor A
MRCA_Brissus_Abatus$factor_A
}
