% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronospace_machinery.R
\name{chronospace}
\alias{chronospace}
\title{Summarize chronospace}
\usage{
chronospace(data_ages)
}
\arguments{
\item{data_ages}{A \code{"nodeAges"} object created using \code{\link[=extract_ages]{extract_ages()}}.}
}
\value{
Percentages of variation in node ages accounted by the factors
considered, both in total and individually, are printed. An object of
class \code{"chronospace"} is returned invisibly, with a list containing
the ordination computed for each factor.
}
\description{
Compute percentages of variation in node ages accounted by each
factor, and extract axes of variation maximizing separation between factor
levels using between-group PCA.
}
\details{
This function summarizes variation in node ages in two ways. First,
a between-groups PCA is performed for each individual factor, extracting
ordination axes that maximize separation between the levels of that factor,
which can be visualized using \code{\link{plot.chronospace}}. Second, a Sum
of squares approaches is used to quantify variation accounted by each
factor, both from the raw total variation and from the bgPC axes computed
for each individual factor.
}
\examples{
#Load ages data
data("data_ages")

#Create chronospace
cspace <- chronospace(data_ages)

#Inspect object
cspace
}
\seealso{
\code{\link{plot.chronospace}}
}
