% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_outputs.R
\name{ltt_sensitivity}
\alias{ltt_sensitivity}
\title{Plot average Lineage Through Time (LTT) curves}
\usage{
ltt_sensitivity(
  data_ages,
  summary = "median",
  colors = 1:5,
  factor = 1:ncol(data_ages$factors),
  plot = T,
  timemarks = NULL,
  gscale = TRUE
)
}
\arguments{
\item{data_ages}{A \code{"nodeAges"} object created using \code{\link[=extract_ages]{extract_ages()}}.}

\item{summary}{Character, indicating whether the 'mean' or 'median' is to be
used in computations.}

\item{colors}{The colors used to represent groups (i.e. levels) of each
factor.}

\item{factor}{Numeric; the factor or factors whose results are to be
plotted and retained. If \code{NULL}, all the factors are retained.}

\item{plot}{whether to plot the results, or only store them (default=TRUE)}

\item{timemarks}{Numeric; an optional vector containing ages to be marked by
vertical lines in LTT plots.}

\item{gscale}{Logical; whether to add chronostratigraphic scale to trees
(via \code{deeptime}).}
}
\description{
For each factor, plot the LTT curve of each level averaged
across the corresponding subsample of chronograms.
}
\examples{
#Load ages data
data("data_ages")

#Create LTT plots
sensiltt <- ltt_sensitivity(data_ages = data_ages, summary = "mean", plot = F)

#Show LTT plot for factor A only
sensiltt$factor_A
}
