% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronospace_machinery.R
\name{print.chronospace}
\alias{print.chronospace}
\title{Print \code{"chronospace"} objects}
\usage{
\method{print}{chronospace}(x)
}
\arguments{
\item{x}{A \code{"chronospace"} object.}
}
\value{
Reports percentages in node ages variation explained by the factors
included.
}
\description{
Print \code{"chronospace"} objects
}
\examples{
#Load ages data
data("data_ages")

#Create chronospace
cspace <- chronospace(data_ages)

#Inspect object
print(cspace)
}
