% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronospace_machinery.R
\name{print.chronospace}
\alias{print.chronospace}
\title{Print \code{"chronospace"} objects}
\usage{
\method{print}{chronospace}(x, ...)
}
\arguments{
\item{x}{A \code{"chronospace"} object.}

\item{...}{Additional arguments passed to \code{print}}
}
\value{
Reports percentages in node ages variation explained by the factors
included.
}
\description{
Print \code{"chronospace"} objects
}
\examples{
#Load ages data
data("echinoid_dates")

#Create chronospace
cspace <- chronospace(echinoid_dates)

#Inspect object
print(cspace)
}
