% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ages.R
\name{print.dataAges}
\alias{print.dataAges}
\title{Print dataAges objects}
\usage{
\method{print}{dataAges}(x)
}
\arguments{
\item{x}{a dataAges object.}
}
\value{

}
\description{
Print dataAges objects
}
