% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_outputs.R
\name{sensitive_nodes}
\alias{sensitive_nodes}
\title{Obtain the most sensitive nodes and depict their age distribution}
\usage{
sensitive_nodes(
  obj,
  amount_of_change,
  chosen_clades,
  factors = 1:length(obj),
  colors = 1:5
)
}
\arguments{
\item{obj}{An object of class \code{"chronospace"} containing one or more
ordinations.}

\item{amount_of_change}{Numeric, specyfing the desired amount of variation in
age (expressed in million of years) above which the nodes are retained and
depicted.}

\item{chosen_clades}{Numeric, indicating the desired number of most sensitive
nodes to be retained and depicted.}

\item{factors}{Numeric; the factor or factors whose results are to be
retained (by default, the first two axes). Ignored if amount_of_change is
explicited.}

\item{colors}{The colors used to represent groups (i.e. levels) of each
factor.}

\item{tree}{An object of class \code{"phylo"} containing the same fixed
topology as the trees from the posterior.}
}
\value{
A panel showing the relative proportions distribution for the age of
each of the most sensitive nodes associated to each factor.
}
\description{
Identify the most sensitive nodes associated to each factor, and
plot their ages proportion distributions.
}
\details{
This function identifies, for each factor, the nodes in the fixed
topology whose ages are most sensitive (i.e. variable) as a function of the
corresponding levels. These can be done either by indicating a threshold
for variation in age above which nodes are considered relevant, or by
specifying the number of most sensitive nodes which must be retained. The
nodes are named using two terminals selected at random from the two
subclades defined by each node.

For each of these most sensitive nodes, the function will plot the
distribution of relative proportion of ages, by level.
}
\references{

}
